+++ lib/executor.dart
@@ -87,7 +87,6 @@ final class Executor {
     List<AccessorListOption>? accessors,
     List<Flag>? flags,
     List<Option>? options,
-    List<SelectedOptions>? selectedOptionses,
     List<String>? defaultCommandPath,
     this.context,
     this.helpFormatter,
@@ -96,7 +95,6 @@ final class Executor {
        accessors = List.unmodifiable(accessors ?? const []),
        flags = List.unmodifiable(flags ?? const []),
        options = List.unmodifiable(options ?? const []),
-       selectedOptionses = List.unmodifiable(selectedOptionses ?? const []),
        defaultCommandPath = defaultCommandPath == null
            ? null
            : List.unmodifiable(defaultCommandPath);
@@ -107,7 +105,6 @@ final class Executor {
   final List<AccessorListOption> accessors;
   final List<Flag> flags;
   final List<Option> options;
-  final List<SelectedOptions> selectedOptionses;
   final List<String>? defaultCommandPath;
   final List<Command> commands;
   final MambaContext? context;
@@ -177,7 +174,6 @@ final class _Execution {
         longDescription: executor.longDescription,
         flags: [...Executor._defaultFlags, ...executor.flags],
         options: executor.options,
-        selectedOptionses: executor.selectedOptionses,
         accessors: executor.accessors,
         commands: executor.commands,
       ),